backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
celery
redis
dnspython
httpx
weasyprint
supabase
stripe
python-dotenv
shodan
groq
beautifulsoup4
Jinja2
